Move sitemap generator command logic to task
