+++ packages/framework/src/Actions/PostBuildTasks/GenerateSitemap.php
@@ -3,7 +3,9 @@
 namespace Hyde\Framework\Actions\PostBuildTasks;
 
 use Hyde\Framework\Contracts\AbstractBuildTask;
-use Illuminate\Support\Facades\Artisan;
+use Hyde\Framework\Helpers\Features;
+use Hyde\Framework\Hyde;
+use Hyde\Framework\Services\SitemapService;
 
 class GenerateSitemap extends AbstractBuildTask
 {
@@ -11,11 +13,38 @@ class GenerateSitemap extends AbstractBuildTask
 
     public function run(): void
     {
-        Artisan::call('build:sitemap');
+        $this->runPreflightCheck();
+
+        file_put_contents(
+            Hyde::getSiteOutputPath('sitemap.xml'),
+            SitemapService::generateSitemap()
+        );
     }
 
     public function then(): void
     {
         $this->writeln("\n".' > Created <info>sitemap.xml</info> in '.$this->getExecutionTime());
+    }
+
+    protected function runPreflightCheck(): bool
+    {
+        if (! Features::sitemap()) {
+            $this->error('Cannot generate sitemap.xml, please check your configuration.');
+
+            if (! Hyde::hasSiteUrl()) {
+                $this->warn('Hint: You don\'t have a site URL configured. Check config/hyde.php');
+            }
+            if (config('site.generate_sitemap', true) !== true) {
+                $this->warn('Hint: You have disabled sitemap generation in config/hyde.php');
+                $this->line(' > You can enable sitemap generation by setting <info>`site.generate_sitemap`</> to <info>`true`</>');
+            }
+            if (! extension_loaded('simplexml') || config('testing.mock_disabled_extensions', false) === true) {
+                $this->warn('Hint: You don\'t have the <info>`simplexml`</> extension installed. Check your PHP installation.');
+            }
+
+            return false;
+        }
+
+        return true;
     }
 }
